pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
]
